refactor(LottoWinningSystem): LottoMachine 이름 수정
전에는 LottoMachine이라는 이름으로 추상적이었습니다
후에는 LottoWiningSystem으로 당첨 시스템이라는 문맥을 추가했습니다
